various: Removes --optimizer option, allows arbitrary specifications.
Functions may be declared like so:
furnace --function-<PREFIX> <ROUTINE>

For example:
furnace --function-vector_render inkscape

This implements autocomplete for these functions also, including any
custom functions declared when in an appropriate folder. Note that the
config and function file must be trusted in this case, as they are run
with no safety.

All routines for functions must now be formatted like so:
__routine__<PREFIX>__<ROUTINE>

For example:
__routine__hash__sha512sum
or
__routine__vector_render__inkscape

This means that no prefix may contain double underscores, but any final
command may do so.

Signed-off-by: William Floyd <william.png2000@gmail.com>