various: Splits function files and removes hard coded file names.
This commit splits all functions from furnace_image_functions.sh and
furnace_functions.sh into topical files in a 'functions' folder. These
file names are not hard coded, and may be swapped in and out at will, so
long as the functions are defined correctly.

In addition to this, furnace_setup.sh automatically appends the correct
bash snippet (temporarily) to export all functions, in much the same way
as the config is pre-processed.

Ultimately, this change should allow more logical modification and
ordering of functions.

Signed-off-by: William Floyd <william.png2000@gmail.com>