render: Lets tsort handle whitespaces correctly, allows optimize only.
Resolves an issue when there is no ending new line in configs or
function files (when sourced by setup).

Having resolved whitespace issues, I have now realised something:
Furnace does not currectly handle editing files in place. For Angl, this
is not an issue - there are no sources that conflict with destination
files. For less automated (that is, optimize only) situations, or cases
when modifying files in place is desirable, there is no working method
as of the moment. Furnace will re-process any in-place file operations,
and I have not even tested self dependency. However, this leads to a
plan for the furture:

Furnace must and shall, in addition to variant work, make a distinction
between source files and produced files, and never the twain shall meet
unsupervised.

The currect plan, as such, will be to implement the following:

Sandboxed processing directories
    Symlink a strict set of dependencies into a temporary directory, and
    extract only the stated file(s) once processed. This ensures that
    any conflicts that occur are strictly due to user error, and that no
    temporary directory be required (though it would be
    prefered, in order to maintain safety). Preference would be to do
    this in memory.

In addition to this, dependencies must be stated as follows.
Source dep:
'./src/assets/minecraft/textures/blocks/clay.svg'
Produced dep:
'./dest/assets/minecraft/textures/blocks/clay.png'

That is, source files will reside in './src/', and produced files will
reside in './dest/'. Furnace must, and shall, warn about any files
that have been found to be changed, produced in, or removed from
'./dest/'. These alterations, should they occur, will be safe from
immediate danger, so long as any run script does not attempt to run in
such as way as to destroy the source files also.

Furnace must, and shall, warn about extraneous files found to be
produced in './dest/' that are not declared in the (planned) PROVIDES
flag (will replace NAME). This warning should be harmless, as these
files are most likely not required. This should be a warning to the user
to use a provided, unique, temorary directory.

In providing both './src/' and './dest/', the user may depend upon both
versions of files that have been modified in place, though this will be
exceedingly rare.

This planned render procedure, when combined with the planned variant
format, shall provide Furnace with the functionality and flexibility
that is desired.

Signed-off-by: William Floyd <william.png2000@gmail.com>